Used the 'FieldArray' component of Redux-Form to have arrays within a form
This is used to create the recipes to match the schema. First the user can enter some general information about the recipe, and then add an indeterminite number of steps to the recipe (fieldarray). up next, getting forms to submit to the DB. TODO: validation and rework other forms to redux-form