Only render the Navbar, before React Rotuer's Switch
Removed Jquery and another script file from the gulp tasks. These were included with the theme and it turns out i dont need any of the functionality that they provide. The JQuery was overriding the Link components behavior and causing a full page refresh when user clicks 'Home' on Navbar. This behavior was breaking SPA rules.. fixed it